OpenCL/CPUContext: use the Common ForkJoinPool instead of a FixedThreadPool
Typically performs better.
